syzkaller/tools/mock_model/requirements.txt
torch>=2.0.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0
